Changed type of batch_indiches to torch.LongTensor. Otherwise trying to index inputs and targets with batch_indices threw a type error on my setup.

+++ Sample Code/classification_test_cpu.py
@@ -188,7 +188,8 @@ def iterate_batches(inputs, targets, batch_size, shuffle=True):
         for start_index in range(0, len(inputs), batch_size):
 
             batch_indices = indices[start_index:start_index + batch_size]
-            batch_indices = torch.from_numpy(batch_indices)
+
+            batch_indices = torch.from_numpy(batch_indices).type(torch.LongTensor)
 
             yield inputs[batch_indices], targets[batch_indices]
